requirements.txt
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
Flask-Cors==3.0.10
python-dotenv==0.19.2
tweepy==4.5.0
python-telegram-bot==13.11
gunicorn==20.1.0
psycopg2-binary==2.9.3
requests==2.27.1
web3==5.28.0
pytest==7.0.1
